+++ jscomp/test/tagged_template_test.js
@@ -13,9 +13,15 @@ function sql(prim0, prim1) {
             ]);
 }
 
+var Pg = {
+  sql: sql
+};
+
 var table = "users";
 
 var id = "5";
+
+var queryWithModule = Tagged_template_libJs.sql`SELECT * FROM ${table} WHERE id = ${id}`;
 
 var query = Tagged_template_libJs.sql`SELECT * FROM ${table} WHERE id = ${id}`;
 
@@ -48,38 +54,50 @@ Mt.from_pair_suites("tagged templates", {
       ],
       tl: {
         hd: [
-          "with externals, it should return the result of the function",
+          "with module scoped externals, it should also return a string with the correct interpolations",
           (function (param) {
               return {
                       TAG: "Eq",
-                      _0: length,
-                      _1: 52
+                      _0: queryWithModule,
+                      _1: "SELECT * FROM 'users' WHERE id = '5'"
                     };
             })
         ],
         tl: {
           hd: [
-            "with rescript function, it should return a string with the correct interpolations",
+            "with externals, it should return the result of the function",
             (function (param) {
                 return {
                         TAG: "Eq",
-                        _0: res,
-                        _1: "| 5 * 10 = 50 |"
+                        _0: length,
+                        _1: 52
                       };
               })
           ],
           tl: {
             hd: [
-              "a template literal tagged with json should generate a regular string interpolation for now",
+              "with rescript function, it should return a string with the correct interpolations",
               (function (param) {
                   return {
                           TAG: "Eq",
-                          _0: "some random " + "string",
-                          _1: "some random string"
+                          _0: res,
+                          _1: "| 5 * 10 = 50 |"
                         };
                 })
             ],
-            tl: /* [] */0
+            tl: {
+              hd: [
+                "a template literal tagged with json should generate a regular string interpolation for now",
+                (function (param) {
+                    return {
+                            TAG: "Eq",
+                            _0: "some random " + "string",
+                            _1: "some random string"
+                          };
+                  })
+              ],
+              tl: /* [] */0
+            }
           }
         }
       }
@@ -87,12 +105,13 @@ Mt.from_pair_suites("tagged templates", {
 
 var extraLength = 10;
 
-exports.sql = sql;
+exports.Pg = Pg;
 exports.table = table;
 exports.id = id;
+exports.queryWithModule = queryWithModule;
 exports.query = query;
 exports.extraLength = extraLength;
 exports.length = length;
 exports.foo = foo;
 exports.res = res;
-/* query Not a pure module */
+/* queryWithModule Not a pure module */