+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 # 11.1.0-rc.3 (Unreleased)
 
+#### :nail_care: Polish
+- No parens around tagged template literals. https://github.com/rescript-lang/rescript-compiler/pull/6639
+
 # 11.1.0-rc.2
 
 #### :rocket: New Feature