+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 #### :nail_care: Polish
 - No parens around tagged template literals. https://github.com/rescript-lang/rescript-compiler/pull/6639
+- Allow identifier with modules in tagged template literals (e.g. Pg.sql`select * from ${table} where id = ${id}`). https://github.com/rescript-lang/rescript-compiler/pull/6645
 
 #### :bug: Bug Fix